fix for /back
For unknown reasons, since 1.21.1, /back started failing for me, from zero problems to not working at all in recent versions.

Debugging it, it seems that the EntityPos is sent as a reference to the playerPreviousPos dictionary, by cloning the entire class using .Copy(), it properly saves the previous player position.